requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.1
langchain-community==0.3.1
langchain-pinecone==0.2.0
langchain-huggingface==0.1.0
pinecone-client==5.0.1
pydantic==2.9.2
python-multipart==0.0.9
sentence-transformers==3.1.1
google-generativeai==0.8.3  
